% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\name{new_cast_cate}
\alias{new_cast_cate}
\title{Create a cast_cate Object}
\usage{
new_cast_cate(effects, variables, n_trees = 1000L)
}
\arguments{
\item{effects}{A \code{data.frame} with columns: \code{lon}, \code{lat}, \code{variable},
\code{cate}.}

\item{variables}{Character vector of variables for which CATE was estimated.}

\item{n_trees}{Integer. Number of causal forest trees used.}
}
\value{
A \code{cast_cate} object.
}
\description{
Create a cast_cate Object
}
\keyword{internal}
