% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\name{new_cast_eval}
\alias{new_cast_eval}
\title{Create a cast_eval Object}
\usage{
new_cast_eval(metrics)
}
\arguments{
\item{metrics}{A \code{data.frame} with columns: \code{model}, \code{auc_mean}, \code{auc_sd},
\code{tss_mean}, \code{tss_sd}.}
}
\value{
A \code{cast_eval} object.
}
\description{
Create a cast_eval Object
}
\keyword{internal}
