% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\name{new_cast_predict}
\alias{new_cast_predict}
\title{Create a cast_predict Object}
\usage{
new_cast_predict(predictions, models)
}
\arguments{
\item{predictions}{A \code{data.frame} with columns: \code{lon}, \code{lat}, and one
\verb{HSS_*} column per model.}

\item{models}{Character vector of model names included.}
}
\value{
A \code{cast_predict} object.
}
\description{
Create a cast_predict Object
}
\keyword{internal}
