% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\name{new_cast_roles}
\alias{new_cast_roles}
\title{Create a cast_roles Object}
\usage{
new_cast_roles(roles)
}
\arguments{
\item{roles}{A \code{data.frame} with columns: \code{variable}, \code{role}, \code{in_degree},
\code{out_degree}, \code{role_score}.}
}
\value{
A \code{cast_roles} object.
}
\description{
Create a cast_roles Object
}
\keyword{internal}
