% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.cast_cate}
\alias{plot.cast_cate}
\title{Plot Spatial CATE Map}
\usage{
\method{plot}{cast_cate}(x, variable = NULL, basemap = "world", ...)
}
\arguments{
\item{x}{A \code{cast_cate} object.}

\item{variable}{Character. Which treatment variable to plot. Default is
the first variable.}

\item{basemap}{Character. Basemap type: \code{"world"}, \code{"china"}, or \code{"none"}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Displays spatially varying Conditional Average Treatment Effects on a
geographic map with a diverging color scale.
}
