% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-curves.R
\name{plot.cast_response}
\alias{plot.cast_response}
\title{Plot Response Curves}
\usage{
\method{plot}{cast_response}(x, ...)
}
\arguments{
\item{x}{A \code{cast_response} object from \code{\link[=cast_response_curves]{cast_response_curves()}}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Univariate curves are drawn per variable (colour = engine, free y
scales); bivariate grids are drawn as raster heatmaps per engine.
}
