% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esm-fit.R
\name{print.cast_esm}
\alias{print.cast_esm}
\title{Print Method for Cast ESM}
\usage{
\method{print}{cast_esm}(x, ...)
}
\arguments{
\item{x}{A \code{cast_esm} object.}

\item{...}{Ignored.}
}
\description{
Print Method for Cast ESM
}
