% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{species_datasets}
\alias{species_datasets}
\alias{Alces_alces}
\alias{Bos_mutus}
\alias{Budorcas_taxicolor}
\alias{Capra_sibirica}
\alias{Capreolus_pygargus}
\alias{Cervus_albirostris}
\alias{Cervus_canadensis}
\alias{Elaphodus_cephalophus}
\alias{Equus_kiang}
\alias{Gazella_subgutturosa}
\alias{Macaca_arctoides}
\alias{Macaca_assamensis}
\alias{Macaca_mulatta}
\alias{Macaca_thibetana}
\alias{Moschus_berezovskii}
\alias{Moschus_chrysogaster}
\alias{Moschus_fuscus}
\alias{Moschus_moschiferus}
\alias{Muntiacus_reevesi}
\alias{Muntiacus_vaginalis}
\alias{Naemorhedus_baileyi}
\alias{Nycticebus_bengalensis}
\alias{Pantholops_hodgsonii}
\alias{Procapra_picticaudata}
\alias{Pseudois_nayaur}
\alias{Rhinopithecus_roxellana}
\alias{Rusa_unicolor}
\alias{SID_14303}
\alias{SID_3814}
\alias{Sus_scrofa}
\alias{Trachypithecus_francoisi}
\title{Bundled Species Occurrence Datasets}
\format{
See \link{Ovis_ammon} for the full column description.
}
\source{
EcoISEA3H project. IUCN Red List spatial data, WorldClim v2,
ENVIREM, SRTM, MODIS land cover.
}
\description{
Occurrence and environmental data for 31 additional mammal species on
the ISEA3H Resolution 9 hexagonal grid covering China. Each dataset
has the same structure as \link{Ovis_ammon}: 3648 rows and 59 columns
including \code{presence}, geographic coordinates, and 50 environmental
variables.

Species: \emph{Alces alces}, \emph{Bos mutus}, \emph{Budorcas taxicolor},
\emph{Capra sibirica}, \emph{Capreolus pygargus}, \emph{Cervus albirostris},
\emph{Cervus canadensis}, \emph{Elaphodus cephalophus}, \emph{Equus kiang},
\emph{Gazella subgutturosa}, \emph{Macaca arctoides}, \emph{Macaca assamensis},
\emph{Macaca mulatta}, \emph{Macaca thibetana}, \emph{Moschus berezovskii},
\emph{Moschus chrysogaster}, \emph{Moschus fuscus}, \emph{Moschus moschiferus},
\emph{Muntiacus reevesi}, \emph{Muntiacus vaginalis}, \emph{Naemorhedus baileyi},
\emph{Nycticebus bengalensis}, \emph{Pantholops hodgsonii},
\emph{Procapra picticaudata}, \emph{Pseudois nayaur},
\emph{Rhinopithecus roxellana}, \emph{Rusa unicolor}, \emph{Sus scrofa},
\emph{Trachypithecus francoisi}, plus two IUCN SID-referenced taxa
(SID 14303, SID 3814).
}
\examples{
data(Alces_alces)
table(Alces_alces$presence)
}
\seealso{
\link{Ovis_ammon}, \link{china_env_grid}
}
