% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\name{new_cast_fit}
\alias{new_cast_fit}
\title{Create a cast_fit Object}
\usage{
new_cast_fit(models, cast_vars, env_vars, scaling, screen = NULL)
}
\arguments{
\item{models}{Named list of fitted model objects.}

\item{cast_vars}{Character vector of variables used for modeling.}

\item{env_vars}{Character vector of all environmental variable names.}

\item{scaling}{List with \code{means} and \code{sds} used for standardization.}

\item{screen}{A \code{cast_select} object (or \code{NULL}).}
}
\value{
A \code{cast_fit} object.
}
\description{
Create a cast_fit Object
}
\keyword{internal}
