% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.cast_ensemble}
\alias{plot.cast_ensemble}
\title{Plot Ensemble Prediction Map}
\usage{
\method{plot}{cast_ensemble}(x, basemap = "world", ...)
}
\arguments{
\item{x}{A \code{cast_ensemble} object.}

\item{basemap}{Character. \code{"world"}, \code{"china"}, or \code{"none"}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Ensemble Prediction Map
}
