% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.cast_necessity}
\alias{plot.cast_necessity}
\title{Plot Knockout Necessity (removed in 0.12.0)}
\usage{
\method{plot}{cast_necessity}(x, ...)
}
\arguments{
\item{x}{Ignored.}

\item{...}{Ignored.}
}
\value{
Never returns; always aborts.
}
\description{
Plot Knockout Necessity (removed in 0.12.0)
}
