% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.cast_project}
\alias{plot.cast_project}
\title{Plot Future Climate Projection}
\usage{
\method{plot}{cast_project}(x, scenario = NULL, basemap = "world", ...)
}
\arguments{
\item{x}{A \code{cast_project} object.}

\item{scenario}{Character. Which scenario to plot. Default is the first.}

\item{basemap}{Character. \code{"world"}, \code{"china"}, or \code{"none"}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object or patchwork composite.
}
\description{
Plot Future Climate Projection
}
